backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.68.1
uvicorn==0.15.0

# Authentication and Security
python-jose[cryptography]==3.3.0
msal==1.17.0
python-multipart==0.0.5

# Exchange Integration
exchangelib==4.7.1
aiohttp==3.8.1

# Data Validation
pydantic==1.8.2
email-validator==1.1.3

# CORS
starlette==0.14.2

# Environment Variables
python-dotenv==0.19.0

# Redis for Caching
redis==4.0.2
aioredis==2.0.0

# Monitoring and Logging
prometheus-client==0.11.0
opentelemetry-api==1.7.1
opentelemetry-sdk==1.7.1
opentelemetry-instrumentation-fastapi==0.26b1
opentelemetry-exporter-prometheus==1.7.1

# Testing
pytest==6.2.5
pytest-asyncio==0.15.1
httpx==0.19.0
pytest-cov==2.12.1

# Development Tools
black==21.7b0
flake8==3.9.2
mypy==0.910
isort==5.9.3
